<feat>(inpersonNFT): set tokenIds as public for setting ticket limit

+++ contracts/ticket.sol
@@ -11,15 +11,30 @@ contract Ticket is Ownable {
     IERC20 public usdc;
     IERC20 public frax;
     IERC20 public dai;
-    mapping (string => uint) public usdTicketPrices;
-    mapping (string => uint) public gohmTicketPrices;
+    mapping(string => uint256) public usdTicketPrices;
+    mapping(string => uint256) public gohmTicketPrices;
     string[] public ticketTypes;
     address public gnosisMultiSigAddr;
     address public inPersonTicketNFTAddr;
-    event Approval(address indexed owner, address indexed spender, uint256 value);
-    event WithdrawFundTo(address indexed gnosisMultiSigAddr, uint256 tokenIdx, uint256 tokenBalance);
+    event Approval(
+        address indexed owner,
+        address indexed spender,
+        uint256 value
+    );
+    event WithdrawFundTo(
+        address indexed gnosisMultiSigAddr,
+        uint256 tokenIdx,
+        uint256 tokenBalance
+    );
 
-    constructor(address multisig, address nftAddr, address gohmAddr, address usdcAddr, address fraxAddr, address daiAddr) {
+    constructor(
+        address multisig,
+        address nftAddr,
+        address gohmAddr,
+        address usdcAddr,
+        address fraxAddr,
+        address daiAddr
+    ) {
         gohm = IERC20(gohmAddr);
         usdc = IERC20(usdcAddr);
         frax = IERC20(fraxAddr);
@@ -29,17 +44,31 @@ contract Ticket is Ownable {
     }
 
     // Modifier to check token allowance
-    modifier checkAllowance(string memory tokenName, string memory ticketName, bool isStableCoin) {
-        uint256 tokenPrice = _getTicketPrice(tokenName, ticketName, isStableCoin);
+    modifier checkAllowance(
+        string memory tokenName,
+        string memory ticketName,
+        bool isStableCoin
+    ) {
+        uint256 tokenPrice = _getTicketPrice(
+            tokenName,
+            ticketName,
+            isStableCoin
+        );
         IERC20 _token = _getTokenIERCbyName(tokenName);
         emit Approval(msg.sender, address(this), tokenPrice);
-        require(_token.allowance(msg.sender, address(this)) >= tokenPrice, "Error");
+        require(
+            _token.allowance(msg.sender, address(this)) >= tokenPrice,
+            "Error"
+        );
         _;
     }
 
-    
-    function setTicketPrice(string memory ticketName, bool isStableCoin, uint ticketPrice) public onlyOwner {
-        if (isStableCoin == true){
+    function setTicketPrice(
+        string memory ticketName,
+        bool isStableCoin,
+        uint256 ticketPrice
+    ) public onlyOwner {
+        if (isStableCoin == true) {
             usdTicketPrices[ticketName] = ticketPrice;
         }
         gohmTicketPrices[ticketName] = ticketPrice;
@@ -49,54 +78,106 @@ contract Ticket is Ownable {
         inPersonTicketNFTAddr = addr;
     }
 
-    function buyTicket(string memory tokenName, string memory ticketName, bool isStableCoin) public checkAllowance(tokenName, ticketName, isStableCoin) {
-        uint256 tokenPrice = _getTicketPrice(tokenName, ticketName, isStableCoin);
+    function buyTicket(
+        string memory tokenName,
+        string memory ticketName,
+        bool isStableCoin
+    ) public checkAllowance(tokenName, ticketName, isStableCoin) {
+        uint256 tokenPrice = _getTicketPrice(
+            tokenName,
+            ticketName,
+            isStableCoin
+        );
         IERC20 token = _getTokenIERCbyName(tokenName);
-        SafeERC20.safeTransferFrom(token, msg.sender, address(this), tokenPrice);
+        SafeERC20.safeTransferFrom(
+            token,
+            msg.sender,
+            address(this),
+            tokenPrice
+        );
         InPersonTicketNFT(inPersonTicketNFTAddr).mintNFT(msg.sender);
     }
-    
+
     function withdrawToken() external onlyOwner {
         // multi-sig: Gnosis wallet address
         IERC20[4] memory tokenArray = [gohm, usdc, frax, dai];
-        for (uint idx=0; idx<tokenArray.length; idx++) {
+        for (uint256 idx = 0; idx < tokenArray.length; idx++) {
             uint256 tokenBalance = tokenArray[idx].balanceOf(address(this));
-            if (tokenBalance != 0){
+            if (tokenBalance != 0) {
                 tokenArray[idx].safeTransfer(gnosisMultiSigAddr, tokenBalance);
                 emit WithdrawFundTo(gnosisMultiSigAddr, idx, tokenBalance);
             }
         }
     }
 
-    function _getTokenIERCbyName(string memory tokenName) private view returns (IERC20){
-        if(keccak256(abi.encodePacked("gohm")) == keccak256(abi.encodePacked(tokenName))) {
+    function _getTokenIERCbyName(string memory tokenName)
+        private
+        view
+        returns (IERC20)
+    {
+        if (
+            keccak256(abi.encodePacked("gohm")) ==
+            keccak256(abi.encodePacked(tokenName))
+        ) {
             return gohm;
-        } else if(keccak256(abi.encodePacked("usdc")) == keccak256(abi.encodePacked(tokenName))){
-            return usdc ;
-        } else if(keccak256(abi.encodePacked("frax")) == keccak256(abi.encodePacked(tokenName))){
-            return frax ;
-        } else if(keccak256(abi.encodePacked("dai")) == keccak256(abi.encodePacked(tokenName))){
-            return dai ;
+        } else if (
+            keccak256(abi.encodePacked("usdc")) ==
+            keccak256(abi.encodePacked(tokenName))
+        ) {
+            return usdc;
+        } else if (
+            keccak256(abi.encodePacked("frax")) ==
+            keccak256(abi.encodePacked(tokenName))
+        ) {
+            return frax;
+        } else if (
+            keccak256(abi.encodePacked("dai")) ==
+            keccak256(abi.encodePacked(tokenName))
+        ) {
+            return dai;
         }
-        revert("Invalid tokenName, it should be one of gohm, usdt, usdc, frax, dai");
+        revert(
+            "Invalid tokenName, it should be one of gohm, usdt, usdc, frax, dai"
+        );
     }
 
-    function _getTicketPrice(string memory tokenName, string memory ticketName, bool isStableCoin) private view returns (uint) {
-        if (isStableCoin == true){
-            return usdTicketPrices[ticketName] * 10 ** _getTokenDecimals(tokenName);
+    function _getTicketPrice(
+        string memory tokenName,
+        string memory ticketName,
+        bool isStableCoin
+    ) private view returns (uint256) {
+        if (isStableCoin == true) {
+            return
+                usdTicketPrices[ticketName] * 10**_getTokenDecimals(tokenName);
         }
-        return gohmTicketPrices[ticketName] * 10 ** _getTokenDecimals(tokenName);
+        return gohmTicketPrices[ticketName] * 10**_getTokenDecimals(tokenName);
     }
 
-    function _getTokenDecimals(string memory tokenName) private pure returns (uint) {
-        if(keccak256(abi.encodePacked("gohm")) == keccak256(abi.encodePacked(tokenName))) {
+    function _getTokenDecimals(string memory tokenName)
+        private
+        pure
+        returns (uint256)
+    {
+        if (
+            keccak256(abi.encodePacked("gohm")) ==
+            keccak256(abi.encodePacked(tokenName))
+        ) {
             return 9;
-        } else if(keccak256(abi.encodePacked("usdc")) == keccak256(abi.encodePacked(tokenName))){
-            return 6 ;
-        } else if(keccak256(abi.encodePacked("frax")) == keccak256(abi.encodePacked(tokenName))){
-            return 18 ;
-        } else if(keccak256(abi.encodePacked("dai")) == keccak256(abi.encodePacked(tokenName))){
-            return 18 ;
+        } else if (
+            keccak256(abi.encodePacked("usdc")) ==
+            keccak256(abi.encodePacked(tokenName))
+        ) {
+            return 6;
+        } else if (
+            keccak256(abi.encodePacked("frax")) ==
+            keccak256(abi.encodePacked(tokenName))
+        ) {
+            return 18;
+        } else if (
+            keccak256(abi.encodePacked("dai")) ==
+            keccak256(abi.encodePacked(tokenName))
+        ) {
+            return 18;
         }
         revert("token Address not found!");
     }
